Added port number as a runtime program option.
Wrote thread safe management of future objects to track errors in requests.

Added interface for how requests will be handled.

Wrote a simple router class: still needs the HttpRequest interface defined so it can grab the first part of the URI.

Added definitions to read and write from the HttpConnection class as well as including an HttpRequest cache variable in the class.

Next step is to implement HttpRequest parsing and using the HttpRequest interface to write the routing function.