Alright, everything I've started to do by using the boost asynchronous http server example is actually kinda solved with the recently accepted beast http library into the boost collection.
So f me.

But in a way this is a good thing.

I'm already using boost asio, and it just is a layer on top of that to implement some of the HTTP protocol. I can focus on some of the more important things to me, including the actual application itself instead of writing code to parse Http requests and serialize Http responses.

The next thing I'm going to try to do is compile time routing resolution. Kinda how the Crow library does it. If I can't reasonably do it, I might just rely on the boost::any class to avoid having to deal with the mess it may become.