Alright,
The scope of this project is getting reigned in. It's primary goal is to be a basic clone of the functionality that roll20.net provides implemented in C++ and having desktop and mobile versions that allow for interconnectedness.

I'm also going to be moving this to gitlab to take advantage of some of their features that require a payment plan on github (my account is grandfathered in to the starter silver tier).